#5: Added a very simple custom flow-analysis
This initial version does not support loops, and instead resets the complete flow-state whenever a future version of flow-state would need to be merged.
It also does not yet support value-dependent flow-state, with VisitCondition() just being a stub that returns the same state for both the true and false branches.