Failing test for deconstructing non-tuple
I think the solution is to locate the Deconstruct method of operation.Value.Type, then create an edge from each parameter, to the lhs of the tuple.
I'm unaware of a Roslyn API method that can get the deconstruct method currently.
Here's some of the internal code related to it: https://github.com/dotnet/roslyn/pull/11873/files
Here's the issue request to expose it on the operation: https://github.com/dotnet/roslyn/issues/33590

For the time-being, it may be worth a simple attempt (find a method member called "Deconstruct" with the right number of parameters), then falling back on just returning rhs?